system/comments_spec.rbの記述
